Rename all party routes to parties
