+++ lib/routes/parties.js
@@ -1,7 +1,7 @@
 const { Router } = require('express');
 const router = Router();
 
-const { setParty, getParty, updateParty } = require('../models/party');
+const { setParty, getParty, updateParty } = require('../models/parties');
 
 router.post('/', async (req, res) => {
   try {
